refactor: moved hardcoded constants to config and refactored part of the codebase to improve structure
