+++ src/handlers/__init__.py
@@ -1,13 +1,25 @@
 from .authorization import authorized_only
+
 from .birthdays import send_birthdays_month, back_to_birthdays
+
 from .commands import send_main_menu, proceed_authorize_users, toggle_admin_mode, temp_authorize_user, \
     send_mass_message,  remind_password, proceed_mass_message, old_button_handler
-# from .departments import ...
+
+from .departments import send_departments, send_inter_department_contacts, send_department_contacts, add_director, \
+    proceed_add_director, send_sub_departments_contacts, send_search_form, proceed_contact_search, \
+    back_to_search_results, manage_additional_departments, manage_sub_department, add_sub_department, \
+    add_sub_department_ans, delete_sub_department
+
 from .employees import add_employee, proceed_add_employee_data, skip_phone, skip_email, skip_username, skip_dob, \
-    send_profile, edit_employee
+    send_profile, edit_employee, new_member_handler, show_keywords, add_keyword, proceed_add_keyword_data, \
+    delete_keyword, confirm_delete_keyword, make_admin, proceed_edit_employee, edit_employee_data_ans, \
+    delete_date_of_birth, delete_employee, confirm_delete_employee
+
 from .google_forms import callback_ans, cancel_form_filling
+
 from .links import send_business_process, send_business_processes_menu, add_link, proceed_add_link_data, send_form, \
     send_helpdesk, show_helpdesk_password, edit_link, proceed_edit_link_data, delete_link_confirmation, delete_link, \
     back_to_send_links
+
 from .main_menu import send_knowledge_base, send_business_processes, send_birthdays, send_contacts_menu, thanks_menu, \
-    send_useful_links, send_form
+    send_useful_links, send_form, back_to_send_contacts_menu
