+++ src/handlers/employees.py
@@ -2,13 +2,18 @@ import asyncio
 import datetime
 import re
 
+from rapidfuzz import process
 from telebot import types
 
-from src.config import bot, process_in_progress, edit_employee_data, authorized_ids, add_employee_data
+from src.config import bot, process_in_progress, edit_employee_data, authorized_ids, add_employee_data, add_keyword_data
 from src.database import DatabaseConnection, update_authorized_users
+from src.handlers import authorized_only
 from src.handlers.authorization import authorized_only
+from src.integrations import proceed_find_user_id, update_employee_in_crm, delete_employee_from_crm, \
+    remove_user_from_chat
 from src.integrations.crm_api_functions import add_employee_to_crm
 from src.integrations.telethon_functions import proceed_find_user_id
+from src.utils import button_names
 
 
 @bot.callback_query_handler(func=lambda call: call.data.startswith('add_employee_'))
@@ -445,3 +450,495 @@ def edit_employee(call):
     if process_in_progress.get(call.message.chat.id) == 'edit_employee':
         del process_in_progress[call.message.chat.id]
         del edit_employee_data[call.from_user.id]
+
+
+@bot.message_handler(content_types=['new_chat_members'])
+def new_member_handler(message):
+    for new_member in message.new_chat_members:
+        if new_member.id == bot.get_me().id:
+            with DatabaseConnection() as (conn, cursor):
+                cursor.execute('INSERT INTO telegram_chats (chat_id, chat_name) VALUES (%s, %s) ',
+                               (message.chat.id, message.chat.title))
+                conn.commit()
+
+
+@bot.callback_query_handler(func=lambda call: call.data.startswith('show_keywords_'))
+@authorized_only(user_type='admins')
+def show_keywords(call):
+    if call.data.startswith('show_keywords_s'):
+        parts = call.data.split('_')
+        search_query = '_'.join(parts[3:-1])
+        employee_id = parts[-1]
+        employee_id = int(employee_id)
+        keyword_btn_callback = f'd_kwd_s_{search_query}_{employee_id}'
+        add_keyword_btn_callback = f'a_kwd_s_{search_query}_{employee_id}'
+        back_btn_callback = f'edit_emp_s_{search_query}_{employee_id}'
+    else:
+        (additional_instance, department_id, intermediate_department_id, sub_department_id,
+         employee_id) = map(int, call.data.split('_')[2:])
+        keyword_btn_callback = (f'd_kwd_{additional_instance}_{department_id}_{intermediate_department_id}_'
+                                f'{sub_department_id}_{employee_id}')
+        add_keyword_btn_callback = (f'a_kwd_{additional_instance}_{department_id}_{intermediate_department_id}_'
+                                    f'{sub_department_id}_{employee_id}')
+        back_btn_callback = f'edit_emp_{additional_instance}_{department_id}_{intermediate_department_id}_' \
+                            f'{sub_department_id}_{employee_id}'
+
+    with DatabaseConnection() as (conn, cursor):
+        cursor.execute('SELECT id, keyword FROM keywords WHERE employee_id = %s ORDER BY keyword', (employee_id,))
+        keywords = cursor.fetchall()
+        cursor.execute('SELECT name FROM employees WHERE id = %s', (employee_id,))
+        employee_name = cursor.fetchone()[0]
+
+    markup = types.InlineKeyboardMarkup(row_width=1)
+    for keyword_id, keyword in keywords:
+        keyword_btn = types.InlineKeyboardButton(text=f'🔍 {keyword}',
+                                                 callback_data=f'{keyword_btn_callback}_{keyword_id}')
+        markup.add(keyword_btn)
+
+    add_keyword_btn = types.InlineKeyboardButton(text='➕ Додати ключове слово', callback_data=add_keyword_btn_callback)
+    back_btn = types.InlineKeyboardButton(text='🔙 Назад', callback_data=back_btn_callback)
+    markup.add(add_keyword_btn, back_btn)
+
+    bot.edit_message_text(f'Ключові слова для контакту <b>{employee_name}</b>:', call.message.chat.id,
+                          call.message.message_id, reply_markup=markup, parse_mode='HTML')
+
+
+@bot.callback_query_handler(func=lambda call: call.data.startswith('a_kwd_'))
+@authorized_only(user_type='admins')
+def add_keyword(call):
+    if call.data.startswith('a_kwd_s'):
+        parts = call.data.split('_')
+        search_query = '_'.join(parts[3:-1])
+        employee_id = parts[-1]
+        employee_id = int(employee_id)
+        back_btn_callback = f'show_keywords_s_{search_query}_{employee_id}'
+    else:
+        (additional_instance, department_id, intermediate_department_id, sub_department_id,
+         employee_id) = map(int, call.data.split('_')[2:])
+        back_btn_callback = f'show_keywords_{additional_instance}_{department_id}_{intermediate_department_id}_' \
+                            f'{sub_department_id}_{employee_id}'
+
+    process_in_progress[call.message.chat.id] = 'add_keyword'
+    add_keyword_data[call.message.chat.id]['employee_id'] = employee_id
+    add_keyword_data[call.message.chat.id]['back_btn_callback'] = back_btn_callback
+
+    cancel_btn = types.InlineKeyboardButton(text='❌ Скасувати', callback_data=back_btn_callback)
+    markup = types.InlineKeyboardMarkup()
+    markup.add(cancel_btn)
+    bot.delete_message(call.message.chat.id, call.message.message_id)
+    sent_message = bot.send_message(call.message.chat.id, '🔍 Введіть ключові слова через кому.\n'
+                                                          'Приклад: <i>програміст, розробник, IT-спеціаліст</i>',
+                                    reply_markup=markup, parse_mode='HTML')
+    add_keyword_data[call.message.chat.id]['saved_message'] = sent_message
+
+
+@bot.message_handler(func=lambda message: message.text not in button_names and process_in_progress.get(
+    message.chat.id) == 'add_keyword')
+@authorized_only(user_type='admins')
+def proceed_add_keyword_data(message):
+    employee_id = add_keyword_data[message.chat.id]['employee_id']
+    back_btn_callback = add_keyword_data[message.chat.id]['back_btn_callback']
+    split_message = message.text.split(',')
+
+    with DatabaseConnection() as (conn, cursor):
+        for keyword in split_message:
+            cursor.execute('INSERT INTO keywords (employee_id, keyword) VALUES (%s, %s)',
+                           (employee_id, keyword.strip()))
+        conn.commit()
+
+    print(f'Keywords "{message.text}" added to employee {employee_id} by {message.from_user.username}.')
+
+    markup = types.InlineKeyboardMarkup()
+    back_btn = types.InlineKeyboardButton(text='🔙 Назад', callback_data=back_btn_callback)
+    markup.add(back_btn)
+
+    bot.delete_message(message.chat.id, add_keyword_data[message.chat.id]['saved_message'].message_id)
+    bot.send_message(message.chat.id, '✅ Ключові слова успішно додані.', reply_markup=markup)
+
+    del process_in_progress[message.chat.id]
+    del add_keyword_data[message.chat.id]
+
+
+@bot.callback_query_handler(func=lambda call: call.data.startswith('d_kwd_'))
+@authorized_only(user_type='admins')
+def delete_keyword(call):
+    if call.data.startswith('d_kwd_s'):
+        parts = call.data.split('_')
+        search_query = '_'.join(parts[3:-2])
+        employee_id = int(parts[-2])
+        keyword_id = int(parts[-1])
+        confirm_delete_keyword_callback = f'cd_kwd_s_{search_query}_{employee_id}_{keyword_id}'
+        back_btn_callback = f'show_keywords_s_{search_query}_{employee_id}'
+    else:
+        (additional_instance, department_id, intermediate_department_id, sub_department_id,
+         employee_id, keyword_id) = map(int, call.data.split('_')[2:])
+        confirm_delete_keyword_callback = (f'cd_kwd_{additional_instance}_{department_id}_{intermediate_department_id}_'
+                                           f'{sub_department_id}_{employee_id}_{keyword_id}')
+        back_btn_callback = (f'show_keywords_{additional_instance}_{department_id}_{intermediate_department_id}_'
+                             f'{sub_department_id}_{employee_id}')
+
+    with DatabaseConnection() as (conn, cursor):
+        cursor.execute('SELECT keyword FROM keywords WHERE id = %s', (keyword_id,))
+        keyword = cursor.fetchone()[0]
+
+    message_text = f'Підтвердіть видалення ключового слова <b>{keyword}</b>:'
+    markup = types.InlineKeyboardMarkup(row_width=1)
+    back_btn = types.InlineKeyboardButton(text='❌ Скасувати видалення', callback_data=back_btn_callback)
+    confirm_btn = types.InlineKeyboardButton(text='✅ Підтвердити видалення',
+                                             callback_data=confirm_delete_keyword_callback)
+    markup.add(confirm_btn, back_btn)
+
+    bot.edit_message_text(message_text, call.message.chat.id, call.message.message_id, reply_markup=markup,
+                          parse_mode='HTML')
+
+
+@bot.callback_query_handler(func=lambda call: call.data.startswith('cd_kwd_'))
+@authorized_only(user_type='admins')
+def confirm_delete_keyword(call):
+    if call.data.startswith('cd_kwd_s'):
+        parts = call.data.split('_')
+        search_query = '_'.join(parts[3:-2])
+        employee_id = int(parts[-2])
+        keyword_id = int(parts[-1])
+        back_btn_callback = f'show_keywords_s_{search_query}_{employee_id}'
+    else:
+        (additional_instance, department_id, intermediate_department_id, sub_department_id,
+         employee_id, keyword_id) = map(int, call.data.split('_')[2:])
+        back_btn_callback = (f'show_keywords_{additional_instance}_{department_id}_{intermediate_department_id}_'
+                             f'{sub_department_id}_{employee_id}')
+
+    with DatabaseConnection() as (conn, cursor):
+        cursor.execute('DELETE FROM keywords WHERE id = %s RETURNING keyword', (keyword_id,))
+        keyword = cursor.fetchone()[0]
+        conn.commit()
+
+    print(f'Keyword "{keyword}" deleted from employee {employee_id} by {call.from_user.username}.')
+
+    message_text = f'✅ Ключове слово <b>{keyword}</b> видалено.'
+    markup = types.InlineKeyboardMarkup()
+    back_btn = types.InlineKeyboardButton(text='🔙 Назад', callback_data=back_btn_callback)
+    markup.add(back_btn)
+
+    bot.edit_message_text(message_text, call.message.chat.id, call.message.message_id, reply_markup=markup,
+                          parse_mode='HTML')
+
+
+@bot.callback_query_handler(func=lambda call: call.data.startswith('make_admin_'))
+@authorized_only(user_type='admins')
+def make_admin(call):
+    employee_id = int(call.data.split('_')[2])
+    with DatabaseConnection() as (conn, cursor):
+        cursor.execute('SELECT id FROM admins WHERE employee_id = %s', (employee_id,))
+        is_admin = cursor.fetchone()
+        if is_admin:
+            cursor.execute('DELETE FROM admins WHERE employee_id = %s', (employee_id,))
+            message_text = f'✅ Користувач {employee_id} більше не є адміністратором.'
+            log_text = f'Employee {employee_id} removed from admins by {call.from_user.username}.'
+        else:
+            cursor.execute('INSERT INTO admins (employee_id) VALUES (%s)', (employee_id,))
+            message_text = f'✅ Користувач {employee_id} тепер є адміністратором.'
+            log_text = f'Employee {employee_id} added to admins by {call.from_user.username}.'
+        conn.commit()
+    print(log_text)
+    update_authorized_users(authorized_ids)
+    bot.delete_message(call.message.chat.id, call.message.message_id)
+    bot.send_message(call.message.chat.id, message_text)
+    bot.send_message(call.message.chat.id, call.message.text, reply_markup=call.message.reply_markup)
+
+
+@bot.callback_query_handler(func=lambda call: call.data.startswith('e_'))
+@authorized_only(user_type='admins')
+def proceed_edit_employee(call):
+    process_in_progress[call.message.chat.id] = 'edit_employee'
+    edit_employee_data[call.from_user.id]['saved_message'] = call.message
+
+    additional_button = None
+
+    if call.data.split('_')[2] == 's':
+        parts = call.data.split('_')
+        search_query = '_'.join(parts[3:-1])
+        employee_id = parts[-1]
+        employee_id = int(employee_id)
+
+        back_btn_callback = f'edit_emp_s_{search_query}_{employee_id}'
+    else:
+        (additional_instance, department_id, intermediate_department_id, sub_department_id,
+         employee_id) = map(int, call.data.split('_')[2:])
+
+        back_btn_callback = (f'edit_emp_{additional_instance}_{department_id}_{intermediate_department_id}_'
+                             f'{sub_department_id}_{employee_id}')
+
+    with DatabaseConnection() as (conn, cursor):
+        cursor.execute('SELECT name FROM employees WHERE id = %s', (employee_id,))
+        employee_name = cursor.fetchone()[0]
+
+    if call.data.startswith('e_name'):
+        edit_employee_data[call.from_user.id]['column'] = ('name', employee_id)
+        message_text = f'✏️ Введіть нове ім\'я для контакту <b>{employee_name}</b>:'
+    elif call.data.startswith('e_phone'):
+        edit_employee_data[call.from_user.id]['column'] = ('phone', employee_id)
+        message_text = f'📞 Введіть новий телефон для контакту <b>{employee_name}</b>:'
+    elif call.data.startswith('e_pos'):
+        edit_employee_data[call.from_user.id]['column'] = ('position', employee_id)
+        message_text = f'💼 Введіть нову посаду для контакту <b>{employee_name}</b>:'
+    elif call.data.startswith('e_uname'):
+        edit_employee_data[call.from_user.id]['column'] = ('telegram_username', employee_id)
+        message_text = f'🆔 Введіть новий юзернейм для контакту <b>{employee_name}</b>:'
+    elif call.data.startswith('e_email'):
+        edit_employee_data[call.from_user.id]['column'] = ('email', employee_id)
+        message_text = f'📧 Введіть новий email для контакту <b>{employee_name}</b>:'
+    elif call.data.startswith('e_dob'):
+        edit_employee_data[call.from_user.id]['column'] = ('date_of_birth', employee_id)
+        message_text = f'🎂 Введіть нову дату народження для контакту <b>{employee_name}</b>:'
+        delete_date_of_birth_btn = types.InlineKeyboardButton(text='🗑️ Видалити дату народження',
+                                                              callback_data=f'del_dob_{employee_id}')
+        additional_button = delete_date_of_birth_btn
+    elif call.data.startswith('e_subdep'):
+        edit_employee_data[call.from_user.id]['column'] = ('sub_department_id', employee_id)
+        message_text = f'🗄️ Введіть приблизну назву відділу для контакту <b>{employee_name}</b>:'
+    else:
+        return
+
+    back_btn = types.InlineKeyboardButton(text='❌ Скасувати', callback_data=back_btn_callback)
+    markup = types.InlineKeyboardMarkup()
+    markup.add(back_btn)
+    if additional_button:
+        markup.add(additional_button)
+    bot.delete_message(call.message.chat.id, call.message.message_id)
+    sent_message = bot.send_message(call.message.chat.id, message_text, reply_markup=markup, parse_mode='HTML')
+    edit_employee_data[call.from_user.id]['saved_markup'] = markup
+    edit_employee_data[call.from_user.id]['saved_message'].message_id = sent_message.message_id
+
+
+@bot.message_handler(func=lambda message: message.text not in button_names and process_in_progress.get(
+    message.chat.id) == 'edit_employee')
+@authorized_only(user_type='admins')
+def edit_employee_data_ans(message):
+    finish_function = True
+    column, employee_id = edit_employee_data[message.chat.id]['column']
+    new_value = message.text
+    with DatabaseConnection() as (conn, cursor):
+        cursor.execute(f'SELECT name FROM employees WHERE id = %s', (employee_id,))
+        employee_data = cursor.fetchone()
+    employee_name = employee_data[0]
+
+    if column == 'name':
+        result_message_text = f'✅ Ім\'я контакту змінено на <b>{new_value}</b>.'
+        log_text = f'Employee {employee_id} name changed to {new_value} by {message.from_user.username}.'
+
+    elif column == 'phone':
+        clear_number = re.match(r'^3?8?(0\d{9})$', re.sub(r'\D', '', new_value))
+        if clear_number:
+            new_value = f'+38{clear_number.group(1)}'
+            result_message_text = f'✅ Номер телефону контакту <b>{employee_name}</b> змінено на <b>{new_value}</b>.'
+            log_text = f'Employee {employee_id} phone changed to {new_value} by {message.from_user.username}.'
+        else:
+            result_message_text = ('🚫 Номер телефону введено невірно.'
+                                   '\nВведіть номер телефону в форматі 0XXXXXXXXX:')
+            log_text = ''
+            finish_function = False
+
+    elif column == 'position':
+        result_message_text = f'✅ Посаду контакту <b>{employee_name}</b> змінено на <b>{new_value}</b>.'
+        log_text = f'Employee {employee_id} position changed to {new_value} by {message.from_user.username}.'
+
+    elif column == 'sub_department_id':
+        with DatabaseConnection() as (conn, cursor):
+            cursor.execute('SELECT id, name FROM sub_departments')
+            sub_departments = cursor.fetchall()
+            original_sub_departments = [(sub_department[0], sub_department[1].strip()) for sub_department in
+                                        sub_departments]
+            sub_departments = [(id, name.lower()) for id, name in original_sub_departments]
+        query = new_value.lower()
+        best_match = process.extractOne(query, [name for id, name in sub_departments])
+        original_best_match = next((id, name) for id, name in original_sub_departments if name.lower() == best_match[0])
+        new_value = original_best_match[0]
+        sub_department_name = original_best_match[1]
+        result_message_text = (f'✅ Відділ контакту <b>{employee_name}</b> змінено на <b>{sub_department_name}</b>.'
+                               f'\nСхожість: {best_match[1]:.1f}%')
+        log_text = f'Employee {employee_id} sub_department_id changed to {new_value} by {message.from_user.username}.'
+
+    elif column == 'telegram_username':
+        searching_message = bot.send_message(message.chat.id, '🔄 Пошук користувача в Telegram...')
+        telegram_user_id = asyncio.run(proceed_find_user_id(new_value))
+        bot.delete_message(message.chat.id, searching_message.message_id)
+        if telegram_user_id is not None:
+            if not new_value.startswith('@'):
+                new_value = f'@{new_value}'
+            update_authorized_users(authorized_ids)
+            result_message_text = f'✅ Юзернейм контакту <b>{employee_name}</b> змінено на <b>{new_value}</b>.'
+            log_text = f'Employee {employee_id} username changed to {new_value} by {message.from_user.username}.'
+        else:
+            result_message_text = (
+                '🚫 Користувач не знайдений. Перевірте правильність введеного юзернейму та спробуйте ще раз.')
+            log_text = ''
+            finish_function = False
+    elif column == 'email':
+        result_message_text = f'✅ Email контакту <b>{employee_name}</b> змінено на <b>{new_value}</b>.'
+        log_text = f'Employee {employee_id} email changed to {new_value} by {message.from_user.username}.'
+    elif column == 'date_of_birth':
+        date_formats = ['%d.%m.%Y', '%d-%m-%Y', '%d/%m/%Y', '%d %m %Y']
+        for date_format in date_formats:
+            try:
+                new_value = datetime.datetime.strptime(new_value, date_format)
+                result_message_text = (f'✅ Дату народження контакту <b>{employee_name}</b> змінено на '
+                                       f'<b>{new_value.strftime("%d/%m/%Y")}</b>.')
+                log_text = (f'Employee {employee_id} date of birth changed to {new_value} by '
+                            f'{message.from_user.username}.')
+                break
+            except ValueError:
+                continue
+        else:
+            result_message_text = ('🚫 Дату народження введено невірно.'
+                                   '\nВведіть дату народження в форматі ДД.ММ.РРРР:')
+            log_text = ''
+            finish_function = False
+    else:
+        return
+
+    saved_message = edit_employee_data[message.chat.id]['saved_message']
+    bot.delete_message(message.chat.id, message.message_id)
+    if edit_employee_data[message.chat.id].get('error_message'):
+        error_message = edit_employee_data[message.chat.id]['error_message']
+        bot.delete_message(message.chat.id, error_message.message_id)
+        del edit_employee_data[message.chat.id]['error_message']
+    else:
+        bot.delete_message(message.chat.id, saved_message.message_id)
+
+    if not finish_function:
+        markup = edit_employee_data[message.chat.id]['saved_markup']
+        error_message = bot.send_message(message.chat.id, result_message_text, reply_markup=markup, parse_mode='HTML')
+        edit_employee_data[message.chat.id]['error_message'] = error_message
+    else:
+        with DatabaseConnection() as (conn, cursor):
+            cursor.execute(f'UPDATE employees SET {column} = %s WHERE id = %s '
+                           f'RETURNING crm_id, name, phone, position, telegram_user_id, telegram_username, email',
+                           (new_value, employee_id))
+            crm_id, name, phone, position, telegram_user_id_crm, telegram_username, email = cursor.fetchone()
+            if column == 'telegram_username':
+                cursor.execute('UPDATE employees SET telegram_user_id = %s WHERE id = %s',
+                               (telegram_user_id, employee_id))
+                print(f'Employee {employee_id} telegram_user_id changed to {telegram_user_id} by '
+                      f'{message.from_user.username}.')
+                telegram_user_id_crm = telegram_user_id
+            conn.commit()
+
+            if column == 'telegram_username':
+                update_authorized_users(authorized_ids)
+
+        update_employee_in_crm(crm_id, name, phone, position, telegram_user_id_crm, telegram_username, email)
+
+        bot.send_message(message.chat.id, result_message_text, parse_mode='HTML')
+        bot.send_message(message.chat.id, text=saved_message.text, reply_markup=saved_message.reply_markup,
+                         parse_mode='HTML')
+
+        del process_in_progress[message.chat.id]
+        del edit_employee_data[message.chat.id]
+        print(log_text)
+
+
+@bot.callback_query_handler(func=lambda call: call.data.startswith('del_dob_'))
+@authorized_only(user_type='admins')
+def delete_date_of_birth(call):
+    employee_id = int(call.data.split('_')[2])
+    with DatabaseConnection() as (conn, cursor):
+        cursor.execute('UPDATE employees SET date_of_birth = NULL WHERE id = %s RETURNING name', (employee_id,))
+        employee_name = cursor.fetchone()[0]
+        conn.commit()
+    markup = call.message.reply_markup
+
+    new_markup = types.InlineKeyboardMarkup()
+    back_button = markup.keyboard[0][0]
+    back_button.text = '🔙 Назад'
+
+    new_markup.add(back_button)
+    bot.edit_message_text(f'✅ Дату народження контакту <b>{employee_name}</b> видалено.', call.message.chat.id,
+                          call.message.message_id, parse_mode='HTML', reply_markup=new_markup)
+    print(f'Employee {employee_id} date of birth deleted by {call.from_user.username}.')
+    del process_in_progress[call.message.chat.id]
+    del edit_employee_data[call.from_user.id]
+
+
+@bot.callback_query_handler(func=lambda call: call.data.startswith('delete_'))
+@authorized_only(user_type='admins')
+def delete_employee(call):
+    if call.data.startswith('delete_s'):
+        search_query, employee_id = call.data.split('_')[2:]
+        employee_id = int(employee_id)
+
+        cancel_btn_callback = f'edit_emp_s_{search_query}_{employee_id}'
+        confirm_btn_callback = f'confirm_delete_s_{employee_id}'
+
+    else:
+        (additional_instance, department_id, intermediate_department_id, sub_department_id,
+         employee_id) = map(int, call.data.split('_')[1:])
+
+        cancel_btn_callback = (f'edit_emp_{additional_instance}_{department_id}_{intermediate_department_id}_'
+                               f'{sub_department_id}_{employee_id}')
+        confirm_btn_callback = (f'confirm_delete_{additional_instance}_{department_id}_{intermediate_department_id}_'
+                                f'{sub_department_id}_{employee_id}')
+
+    cancel_btn = types.InlineKeyboardButton(text='❌ Скасувати видалення', callback_data=cancel_btn_callback)
+    confirm_btn = types.InlineKeyboardButton(text='✅ Підтвердити видалення', callback_data=confirm_btn_callback)
+    markup = types.InlineKeyboardMarkup(row_width=1)
+    markup.add(confirm_btn, cancel_btn)
+
+    with DatabaseConnection() as (conn, cursor):
+        cursor.execute('SELECT name FROM employees WHERE id = %s', (employee_id,))
+        employee_name = cursor.fetchone()[0]
+
+    bot.edit_message_text(f'Ви впевнені, що хочете видалити контакт <b>{employee_name}</b>?', call.message.chat.id,
+                          call.message.message_id, reply_markup=markup, parse_mode='HTML')
+
+
+@bot.callback_query_handler(func=lambda call: call.data.startswith('confirm_delete_'))
+@authorized_only(user_type='admins')
+def confirm_delete_employee(call):
+    if call.data.startswith('confirm_delete_s'):
+        employee_id = int(call.data.split('_')[3])
+
+        back_btn_callback = 'back_to_send_contacts'
+
+    else:
+        (additional_instance, department_id, intermediate_department_id, sub_department_id,
+         employee_id) = map(int, call.data.split('_')[2:])
+
+        back_btn_callback = (f'sub_dep_{additional_instance}_{department_id}_{intermediate_department_id}_'
+                             f'{sub_department_id}')
+
+    back_btn = types.InlineKeyboardButton(text='🔙 Назад', callback_data=back_btn_callback)
+    markup = types.InlineKeyboardMarkup()
+    markup.add(back_btn)
+
+    with DatabaseConnection() as (conn, cursor):
+        cursor.execute('SELECT name, telegram_user_id FROM employees WHERE id = %s', (employee_id,))
+        employee_name, telegram_user_id = cursor.fetchone()
+        cursor.execute('DELETE FROM employees WHERE id = %s RETURNING crm_id', (employee_id,))
+        crm_id = cursor.fetchone()[0]
+        conn.commit()
+        cursor.execute('SELECT chat_id, chat_name from telegram_chats')
+        chats = cursor.fetchall()
+
+    delete_employee_from_crm(crm_id)
+
+    print(f'Employee {employee_name} deleted by {call.from_user.username}.')
+    update_authorized_users(authorized_ids)
+
+    successful_chats = []
+
+    for chat_id, chat_name in chats:
+        try:
+            remove_user_from_chat(bot, chat_id, telegram_user_id)
+        except Exception as e:
+            print(f'Error while removing user from chat: {e}')
+            continue
+        successful_chats.append(chat_name)
+
+    message = f'✅ Контакт <b>{employee_name}</b> видалено.'
+
+    if successful_chats:
+        chat_list = ', '.join(successful_chats)
+        message += f'\n\nКонтакт також було видалено з чатів: <b>{chat_list}</b>.'
+
+    bot.edit_message_text(message, call.message.chat.id,
+                          call.message.message_id, parse_mode='HTML', reply_markup=markup)
